Uprawnienia grupowe
Dodano sprawdzanie przynależności do grupy i przyznawania dostępu na tej podstawie do edycji książek. Wypełniono bazę danych.

+++ myapp/views.py
@@ -14,31 +14,26 @@ from django.db.models import Q
 def login_user(request, format=None):
     password = request.data.get('password')
     username_or_email = request.data.get('login')
-    # Sprawdzenie czy użytkownik istnieje w bazie danych
     user = None
     if username_or_email and '@' in username_or_email:
-
         user = User.objects.filter(email=username_or_email).first()
     elif username_or_email:
-
         user = User.objects.filter(username=username_or_email).first()
-
-    # Jeśli użytkownik istnieje, sprawdź poprawność hasła
     if user and user.check_password(password):
         token = Token.objects.get_or_create(user=user)[0]
-        response = Response({'message': 'Zalogowano pomyślnie'})
-        response.set_cookie('auth_token', token.key)  # ustawienie ciasteczka
+        response = Response({'message': f'Login succesfull {user.username}'})
+        response.set_cookie('auth_token', token.key)
         response['Authorization'] =  'Token ' + token.key
         return response
     else:
-        return Response({'error': 'Nieprawidłowy login lub hasło'})
+        return Response({'error': 'Incorrect password'})
 
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
 def logout_user(request, format=None):
     request.user.auth_token.delete()
-    response = Response({'message': 'Wylogowano pomyślnie'})
-    response.delete_cookie('auth_token')  # usuwanie ciasteczka
+    response = Response({'message': 'Logout succesfull'})
+    response.delete_cookie('auth_token')
     return response
 
 @api_view(['POST'])
@@ -47,15 +42,16 @@ def register_user(request, format=None):
         serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'message': 'Register succesfull'},serializer.data, status=status.HTTP_201_CREATED)
+        return Response({'message': 'Register unsuccesfull'},serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET'])
 def category_list(request,format=None):
     if request.method == 'GET':
-        categories =  Category.objects.all()
+        categories = Category.objects.all()
         serializer = CategorySerializer(categories, many=True)
         return Response(serializer.data)
+
 @api_view(['GET', 'POST'])
 def books_list(request, format=None):
     if request.method == 'GET':
@@ -70,24 +66,23 @@ def books_list(request, format=None):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-@api_view(['GET', 'POST'])
+@api_view(['GET'])
 def review_list(request, format=None):
     if request.method == 'GET':
         reviews = Review.objects.all()
         serializer = ReviewSerializerSmall(reviews, many=True)
         return Response(serializer.data)
-    elif request.method == 'POST':
-        book_title = request.data.get('book_title')
-
-        serializer = ReviewSerializer(data=request.data)
-        if serializer.is_valid():
-            try:
-                book = Book.objects.get(title=book_title)
-            except Book.DoesNotExist:
-                return Response({'error': 'Book does not exist'}, status=status.HTTP_404_NOT_FOUND)
-            serializer.save(book=book)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    # elif request.method == 'POST':
+    #     book_title = request.data.get('book_title')
+    #     serializer = ReviewSerializer(data=request.data)
+    #     if serializer.is_valid():
+    #         try:
+    #             book = Book.objects.get(title=book_title)
+    #         except Book.DoesNotExist:
+    #             return Response({'error': 'Book does not exist'}, status=status.HTTP_404_NOT_FOUND)
+    #         serializer.save(book=book)
+    #         return Response(serializer.data, status=status.HTTP_201_CREATED)
+    #     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 
@@ -98,54 +93,65 @@ def book_detail(request, id, format=None):
         book = Book.objects.get(pk=id)
     except Book.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
-
     if request.method == 'GET':
         serializer = BookSerializer(book)
         return Response(serializer.data)
+
     elif request.method == 'PUT':
+        if not request.user.groups.filter(name="bookadmin").exists():
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         serializer = BookSerializer(book, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     elif request.method == 'DELETE':
+        if not request.user.groups.filter(name="bookadmin").exists():
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         book.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-    elif request.method == 'POST':
-        existing_review = Review.objects.filter(book=id, user=request.user)
-        if existing_review.exists():
-            return Response({'error': 'User already reviewed this book.'}, status=status.HTTP_400_BAD_REQUEST)
 
-        serializer = ReviewSerializer(data=request.data,context={'request': request})
-        if serializer.is_valid():
-            serializer.save(book=book, user=request.user)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['GET', 'PUT', 'DELETE'])
+
+
+@api_view(['GET', 'PUT', 'DELETE','POST'])
 def review_detail(request, id, format=None):
+    if request.method == 'POST':
+        try:
+            book = Book.objects.get(pk=id)
+        except Review.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        if request.user.is_authenticated:
+            existing_review = Review.objects.filter(book=id, user=request.user)
+            if existing_review.exists():
+                return Response({'error': 'User already reviewed this book.'}, status=status.HTTP_400_BAD_REQUEST)
+            serializer = ReviewSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save(book=book, user=request.user)
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response({'error': 'User is not authenticated.'}, status=status.HTTP_401_UNAUTHORIZED)
     try:
         review = Review.objects.get(pk=id)
     except Review.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
-
     if request.method == 'GET':
         serializer = ReviewSerializer(review)
         return Response(serializer.data)
-
     elif request.method == 'PUT':
         if review.user == request.user:
-            serializer = ReviewSerializer(review, data=request.data,context={'request': request})
+            serializer = ReviewSerializer(review, data=request.data,context={'request': request, 'flag': True})
             if serializer.is_valid():
                 serializer.save()
                 return Response(serializer.data)
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        return Response(status=status.HTTP_401_UNAUTHORIZED)
+        return Response({'error': 'You can only modify your reviews'},status=status.HTTP_401_UNAUTHORIZED,)
     elif request.method == 'DELETE':
         if review.user == request.user:
             review.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
-        return Response(status=status.HTTP_401_UNAUTHORIZED)
+        return Response({'error': 'You can only delete your reviews'},status=status.HTTP_401_UNAUTHORIZED)
 
 @api_view(['GET'])
 def search_book(request):
